Make sure we have set_count and set_count_paged set before trying to use them
If an option hasn't yet been set for a custom post type, or if an option is set to 0, then the keys for set_count and set_count_paged are never set in our final options property. Rather than try to use them anyway, we should just return found posts as we left it.